configure models and DTOs & add logback xml

+++ backend/src/main/java/com/softeng/backend/model/BudgetByCostType.java
@@ -1,6 +1,7 @@
 package com.softeng.backend.model;
 
 import jakarta.persistence.*;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.Data;
 
 import java.math.BigDecimal;
@@ -11,7 +12,7 @@ import java.math.BigDecimal;
 public class BudgetByCostType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(unique = true, nullable = false)
     private String typeName;
